chore(rpsbml.py): remove score computation from __init__
Score is related to rpSBML objects only. 'inFile' parameter is in most cases pure SBML that does not
contain 'rp_pathway' field which is needed in compte_score() method. This method is now called from
outside of rpSBML constructor when we are sure that rpSBML is a real rpSBML object (like in
rpCompletion for instance). In addition, compute_sore() method now returns the score value.